requirements.txt
django==2.2
djangorestframework==3.9.2
PyJWT==1.7.1
